moving files to correct folder and removing program.cs by making csproj a library
