+++ src/Cli/Abstractions/Environment/EnvironmentProvider.cs
@@ -4,6 +4,7 @@
 // Note that this file is copied from: https://github.com/dotnet/sdk/blob/4a81a96a9f1bd661592975c8269e078f6e3f18c9/src/Cli/Microsoft.DotNet.Cli.Utils/EnvironmentProvider.cs
 // Once the dotnet cli utils package is in a published consumable state, we will migrate over to use that
 
+using System;
 using Azure.Functions.Cli.Abstractions.Extensions;
 using System.Runtime.InteropServices;
