+++ src/Cli/Abstractions/Logging/AnsiConsole.cs
@@ -4,7 +4,7 @@
 // Note that this file is copied from: https://github.com/dotnet/sdk/blob/4a81a96a9f1bd661592975c8269e078f6e3f18c9/src/Cli/Microsoft.DotNet.Cli.Utils/AnsiConsole.cs
 // Once the dotnet cli utils package is in a published consumable state, we will migrate over to use that
 
-namespace Azure.Functions.Cli.Abstractions
+namespace Azure.Functions.Cli.Abstractions.Logging
 {
     public class AnsiConsole
     {
@@ -60,7 +60,7 @@ namespace Azure.Functions.Cli.Abstractions
             }
 
             _boldRecursion += bold ? 1 : -1;
-            if (_boldRecursion > 1 || (_boldRecursion == 1 && !bold))
+            if (_boldRecursion > 1 || _boldRecursion == 1 && !bold)
             {
                 return;
             }
