--- src/Cli/Abstractions/Program.cs
@@ -1,12 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-namespace Azure.Functions.Cli.Abstractions
-{
-    internal class Program
-    {
-        internal static void Main(string[] args)
-        {
-        }
-    }
-}
